FEAT final

+++ routes/cinema.routes.js
@@ -1,10 +1,10 @@
 const express = require('express');
-const router = express.Router();
+const cinemasRouter = express.Router();
 const createError = require('../utils/errors/create-error.js');
 
 const Cinema = require('../models/Cinema.js');
 
-router.get('/', async (req, res, next) => {
+cinemasRouter.get('/', async (req, res, next) => {
 	try {
 		const cinemas = await Cinema.find().populate('movies');
 		return res.status(200).json(cinemas)
@@ -13,13 +13,17 @@ router.get('/', async (req, res, next) => {
 	}
 });
 
-router.post('/create', async (req, res, next) => {
+cinemasRouter.post('/create', async (req, res, next) => {
     try{
         const newCinema = new Cinema({
             name: req.body.name,
             location: req.body.location,
             movies: []
         });
+        if (!newCinema) {       
+            return next(createError('No se ha podido crear el cine', 404));
+        }
+
 
         const createdCinema = await newCinema.save();
         return res.status(201).json(createdCinema);
@@ -28,9 +32,12 @@ router.post('/create', async (req, res, next) => {
     }
 });
 
-router.delete('/:id', async (req, res, next) => {
+cinemasRouter.delete('/:id', async (req, res, next) => {
     try{
         const{id} = req.params;
+        if(!id){
+            return next(createError('No se ha podido eliminar el cine', 404));
+        }
         const cinemaDeleted = await Cinema.findByIdAndDelete(id);
         return res.status(200).json(cinemaDeleted);
     }catch(err){
@@ -38,10 +45,15 @@ router.delete('/:id', async (req, res, next) => {
     }
 });
 
-router.put('/add-movie', async (req, res, next) => {
+cinemasRouter.put('/add-movie', async (req, res, next) => {
     try {
         const { cinemaId } = req.body;
         const { movieId } = req.body;
+        if (!cinemaId || !movieId) {
+            return next(createError('No se ha podido añadir la película', 404));
+        }
+        const cinema = await Cinema
+        
         const updatedCinema = await Cinema.findByIdAndUpdate(
             cinemaId,
             { $push: { movies: movieId } },
@@ -49,8 +61,9 @@ router.put('/add-movie', async (req, res, next) => {
         );
         return res.status(200).json(updatedCinema);
     } catch (err) {
+
         return next(err);
     }
 });
 
-module.exports = router;
+module.exports = cinemasRouter;
